fix: property getters are broken
The generated property getters are sometimes broken and mismatch names and/or types of the json2jsii generated type.

Since they are non-essential, remove them for now until we have a way to reflect on type generated by json2jsii in order to propertly produce these properties.

Add an example (`aqua-enterprise-enforcer.json`) which demonstrates this case.

Added a jsii compilation stage to the tests so that we can fully validate that the generated output can be compiled.