fix: property getters are broken so expose "props" instead (#57)
The generated property getters are sometimes broken and mismatch names and/or types of the json2jsii generated type. Instead, just expose `props` which returns the properties passed to the constructor.

Add an example (`aqua-enterprise-enforcer.json`) which demonstrates this case.

Added a jsii compilation stage to the tests so that we can fully validate that the generated output can be compiled.

BREAKING CHANGE: the getters for resource properties are no longer available on the generated type. Instead, use `resource.props.xxx` to access the initialization values.