feat: add validation for supported synthesizers in sc codegen (#224)
_This does not add any active code paths._

For cdk v2, using a service catalog provisioning construct does not
make sense for the standard case with the DefaultStackSynthesizer as the cdk bootstrap roles
will invoke Service Catalog to provision the product and will not have
the correct permissions or associations. This change will add that validation
to the codegen so it can fail at synth time instead of during deployment.

Co-authored-by: Aidan Crank <arcrank@gmail.com>