feat: enable sc command for cdk-import (#226)
Add entry-point and cli commands for active service catalog import path.

This change adds a subcommand for sc as well as a default path for cfn to maintain consistency.

The sc subcommand supports pulling in products based on a default strategy or by pulling in a product directly by IDs

Co-authored-by: Aidan Crank <arcrank@gmail.com>