chore: Add and upgrade dependencies
We need `aws-cdk-lib` and `constructs` in our codegen files because we are rendering
an L2 that has an L1 resource in it.

Due to jsii packaging we need these in dependencies otherwise a fresh install of the tool will
fail if you do not have aws-cdk-lib installed.

Also added blip on configuring `AWS_REGION`

Add small python example for SC